resolve circular dependencies in gestureHandlerCommon
