+++ src/handlers/gestures/GestureDetector/attachHandlers.ts
@@ -5,7 +5,7 @@ import RNGestureHandlerModule from '../../../RNGestureHandlerModule';
 import {
   filterConfig,
   scheduleFlushOperations,
-} from '../../gestureHandlerCommon';
+} from 'src/handlers/gestureCommonUtils';
 import { ComposedGesture } from '../gestureComposition';
 import { ActionType } from '../../../ActionType';
 import { Platform } from 'react-native';
